Change file organization(OrderBook)
