+++ src/session/runner.rs
@@ -8,7 +8,8 @@ use pyo3::{pyclass, pymethods, Py, PyAny, PyErr, PyObject, Python};
 use rusqlite::ffi::SQLITE_FCNTL_CKSM_FILE;
 use rust_decimal::prelude::ToPrimitive;
 
-use crate::common::{MarketMessage, MarketStream, MicroSec, Order, Trade};
+use crate::common::{MarketMessage, MarketStream, MicroSec, Order, Trade, MarketConfig};
+use crate::exchange::binance::config;
 
 use super::{has_method, Session};
 
@@ -45,8 +46,12 @@ impl Runner {
         let stream = stream.reciver;
 
         let result = Python::with_gil(|py| {
+            let config = market.getattr(py, "market_config").unwrap();
+            let config = config.extract::<MarketConfig>(py).unwrap();
             // TODO: implement reflet session name based on agent name
-            let py_session = Py::new(py, Session::new(market, false, None)).unwrap();
+            let mut session = Session::new(market, false, &config, None);
+            session.open_log(&session.session_name.clone());
+            let py_session = Py::new(py, session).unwrap();
 
             // TODO: implment on_clock;
             let has_on_clock = has_method(agent, "on_clock");
