+++ src/session/session.rs
@@ -1,4 +1,6 @@
-use numpy::PyArray2;
+use std::{fs::OpenOptions, path::Path};
+use std::io::Write;
+
 use pyo3::{pyclass, pymethods, types::PyTuple, PyAny, PyObject, Python};
 use pyo3_polars::PyDataFrame;
 use rust_decimal::{prelude::ToPrimitive, Decimal};
@@ -7,7 +9,7 @@ use rust_decimal_macros::dec;
 use crate::{
     common::{
         date_string, hour_string, min_string, AccountStatus, MarketStream, MicroSec, OrderSide,
-        OrderStatus, NOW, PRICE_SCALE, SIZE_SCALE,
+        OrderStatus, NOW, MarketConfig,
     },
     exchange::binance::Market,
 };
@@ -19,6 +21,8 @@ use crate::common::MarketMessage;
 use crate::common::Order;
 use crate::common::Trade;
 
+
+
 #[pyclass(name = "Session")]
 #[derive(Debug)]
 pub struct Session {
@@ -28,7 +32,7 @@ pub struct Session {
     market: PyObject,
     current_time: MicroSec,
     dummy: bool,
-    session_name: String,
+    pub session_name: String,
     order_number: i64,
 
     commission_home: Decimal,
@@ -39,12 +43,16 @@ pub struct Session {
     free_foreign_change: Decimal,
     lock_home_change: Decimal,
     lock_foreign_change: Decimal,
+
+    market_config: MarketConfig,
+
+    log_file: Option<std::fs::File>,
 }
 
 #[pymethods]
 impl Session {
     #[new]
-    pub fn new(market: PyObject, dummy: bool, session_name: Option<&str>) -> Self {
+    pub fn new(market: PyObject, dummy: bool, market_config: &MarketConfig, session_name: Option<&str>) -> Self {
         let session_name = match session_name {
             Some(name) => name.to_string(),
             None => {
@@ -77,6 +85,10 @@ impl Session {
             free_foreign_change: dec![0.0],
             lock_home_change: dec![0.0],
             lock_foreign_change: dec![0.0],
+
+            market_config: market_config.clone(),
+
+            log_file:None,
         };
 
         session.load_order_list().unwrap();
@@ -190,11 +202,14 @@ impl Session {
         })
     }
 
-    // TODO: Decimal
+
     pub fn market_order(&mut self, side: OrderSide, size: Decimal) -> Result<Py<PyAny>, PyErr> {
-        let size = size.round_dp(SIZE_SCALE);
+        let size_scale = self.market_config.size_scale;
+        let size = size.round_dp(size_scale);
 
         let local_id = self.new_order_id(&side);
+
+        log::debug!("market_order: side={:?}, size={}", side, size);
 
         Python::with_gil(|py| {
             self.market
@@ -202,24 +217,34 @@ impl Session {
         })
     }
 
-    // TODO: Decimal
+
     pub fn limit_order(
         &mut self,
         side: OrderSide,
-        size: Decimal,
         price: Decimal,
+        size: Decimal,
     ) -> Result<Vec<Order>, PyErr> {
-        let price = price.round_dp(PRICE_SCALE);
-        let size = size.round_dp(SIZE_SCALE);
+        let price_scale = self.market_config.price_scale;
+        let pricedp = price.round_dp(price_scale);
+
+        let size_scale = self.market_config.size_scale;        
+        let sizedp = size.round_dp(size_scale);
 
         // first push order to order list
         let local_id = self.new_order_id(&side);
+
+        log::debug!(
+            "limit_order: side={:?}, size={}, price={}",
+            side,
+            sizedp,
+            pricedp
+        );
 
         // then call market.limit_order
         let r = Python::with_gil(|py| {
             let result = self
                 .market
-                .call_method1(py, "limit_order", (side, size, price, local_id));
+                .call_method1(py, "limit_order", (side, pricedp, sizedp, local_id));
 
             match result {
                 // if success update order list
@@ -239,7 +264,7 @@ impl Session {
                     return Ok(orders);
                 }
                 Err(e) => {
-                    log::error!("limit_order error: {:?}", e);
+                    log::error!("limit_order error: {:?}  / priceorg={:?}/sizeorg={:?}", e, price, size);
                     return Err(e);
                 }
             }
@@ -332,7 +357,9 @@ impl Session {
             log::error!("Unknown order side: {:?}", order.order_side)
         }
 
-        self.update_lalance(order);
+        self.update_balance(order);
+
+        self.log(&order.__str__());
     }
 
     fn new_order_id(&mut self, side: &OrderSide) -> String {
@@ -383,7 +410,7 @@ impl Session {
         return r;
     }
 
-    pub fn update_lalance(&mut self, order: &Order) {
+    pub fn update_balance(&mut self, order: &Order) {
         self.commission_foreign += order.commission_foreign;
         self.commission_home += order.commission_home;
 
@@ -395,4 +422,26 @@ impl Session {
         self.free_foreign_change += order.free_foreign_change;
         self.lock_foreign_change += order.lock_foreign_change;
     }
+
+    pub fn open_log(&mut self, path: &str) -> Result<(), std::io::Error> {
+        let log_file = Path::new(path).with_extension("log");
+
+        self.log_file = Some(
+            OpenOptions::new()
+                .append(true)
+                .create(true)
+                .open(log_file)?
+        );
+        Ok(())
+    }
+
+    pub fn log(&mut self, message: &str) -> Result<(), std::io::Error> {
+        if let Some(file) = &mut self.log_file {
+            writeln!(file, "{}", message)?;
+            //file.write_all(format!("{}", message)?.as_bytes())?;
+        }
+
+        Ok(())
+    }
+
 }
