+++ src/test/binance_test/mmbot.py
@@ -13,10 +13,10 @@ from time import sleep
 import polars as pl
 
 ORDER_SIZE = 0.001
-SPREAD_TRIGGER = 0.0002
-SPREAD_UPDATE =  0.0001
+SPREAD_TRIGGER = 0.0004
+CLOSE_TRIGGER =  0.0003
 PRICE_DIFF = 0.01
-SKIP_SIZE = 0.005
+SKIP_SIZE = 0.025
 
 class MyAgent:
     def __init__(self):
@@ -31,7 +31,8 @@ class MyAgent:
         ask = session.asks.filter(pl.col('size') > skip_size).head()['price'][0]
         
         return bid, ask, ask - bid
-    
+
+    """    
     def on_tick(self, session, side, price, size):
         # スプレットを確認
         bid, ask, spread = self.calc_spread(session)
@@ -41,29 +42,27 @@ class MyAgent:
         amount_buy = session.buy_order_amount
         amount_sell = session.sell_order_amount
 
-        if amount_buy != 0 and amount_sell == 0:  # 買いオーダーだけ残っている場合
-                # 買いオーダーの乖離をみて、離れていれば売り
-                buy_price = session.buy_orders[0].order_price
-                if bid - buy_price > price * SPREAD_UPDATE:
-                    order_size = ORDER_SIZE - session.buy_orders[0].remain_size
-                    session.cancel_order(session.buy_orders[0].order_id)
-                    session.market_order(OrderSide.Buy, order_size)
-               
-        if amount_sell != 0 and amount_buy == 0: #　売りオーダーだけ残っている場合
-                # 売りオーダーの乖離をみて離れていれば修正            
-                sell_price = session.sell_orders[0].order_price    
-                if sell_price - ask > price * SPREAD_UPDATE:
-                    order_size = ORDER_SIZE - session.sell_orders[0].remain_size
-                    session.cancel_order(session.sell_orders[0].order_id)
-                    session.limit_order(OrderSide.Sell, ask + PRICE_DIFF, order_size)
-        
-        # スプレッドが大きい場合
-        if spread > price * SPREAD_TRIGGER:
-            # まだ指値が両サイドに入っていない場合は、両サイドに指値を入れる。
-            if amount_buy == 0 and amount_sell == 0:
+        if amount_buy == 0 and amount_sell == 0:
+            # スプレッドが大きい場合
+            if spread > price * SPREAD_TRIGGER:
+                # まだ指値が両サイドに入っていない場合は、両サイドに指値を入れる。
                 session.limit_order(OrderSide.Buy, bid + PRICE_DIFF, ORDER_SIZE)
                 session.limit_order(OrderSide.Sell, ask - PRICE_DIFF, ORDER_SIZE)
-
+        elif amount_buy != 0 and amount_sell == 0:  # 買いオーダーだけ残っている場合
+                # 買いオーダーの乖離をみて、離れていれば売り
+                buy_price = session.buy_orders[0].order_price
+                if bid - buy_price > price * CLOSE_TRIGGER:
+                    order_size = session.buy_orders[0].remain_size
+                    session.cancel_order(session.buy_orders[0].order_id)
+                    session.market_order(OrderSide.Buy, order_size)
+        elif amount_sell != 0 and amount_buy == 0: #　売りオーダーだけ残っている場合
+                # 売りオーダーの乖離をみて離れていれば修正            
+                sell_price = session.sell_orders[0].order_price    
+                if sell_price - ask > price * CLOSE_TRIGGER:
+                    order_size = session.sell_orders[0].remain_size
+                    session.cancel_order(session.sell_orders[0].order_id)
+                    session.market_order(OrderSide.Sell, order_size)
+    """
 
     def on_update(self, session, updated_order):
         self.home += updated_order.home_change
@@ -82,9 +81,14 @@ class MyAgent:
         print("TOTAL ASSETS: ", account.home + account.foreign * session.bids[0]['price'][0])
 
     
+
+init_log()
+    
 market = BinanceMarket(BinanceConfig.TEST_BTCUSDT)
 
 market.cancel_all_orders()
+
+
 
 print(BinanceConfig.TEST_BTCUSDT)
 
@@ -94,7 +98,7 @@ market.start_user_stream()
 agent = MyAgent()
 runner = Runner()
 
-
+#init_debug_log()
 
 
 runner.run(market, agent)
